Big update
Added support to change (most) previously hardcoded values in the Compatibillity Adapter, the first step to support 1.15 again.
It also improves general code quality.

There is a known bug that causes a fatal memory leak with the fire enchantment, while I assume the changes are not the cause of this, this should be investigated ASAP