Fix OOM with the Fire enchantment
Seems like he actually was saying the truth.

Anyways, I got no idea why I went to the caching route because the issue was something else,, but I doubt that it is much harmfull to do that, so now it is implemented.
For anyone reading this commit, I advice to take the dangers of the while loop pretty seriously, as otherwise these bugs happen