Highly experimental optimisations
Basically the Lumber, Shred, Mow, Pierce, Extraction and Plough enchantments are no longer prioritized, so there is going to be an issue with those enchantments, if the code I removed existed for a reason.

I talked about this issue with the maintainer of the upstream repo, and he doesn't seem to exactly know to why as the code existed in the first place and according to him was used to shuffle enchantments with the anvil, which failed because it was a treemap (there's no treemap) that sorted itself automatically.

The exact reasoning behind the code is unknown and so are the long-term effects, but removing it would mean that up to 2520 comparisions and 36 LinkedHashMap.contains() operations (along other things) are saved every tick for every player