Refractored Enchantment merging
And even more temporary bloat added; it will feel good to finally remove it all in the v3.0.0 refactor. Aside of that I feel like I'm doing a little bit overboard with lambdas, however they should still be faster than whatever we had before and should be easier to maintain (or more specifically, won't need maintainance at all). Well the last part may be a bit wrong, but as soon as we can get the Enchantments to be Enums instead of whatever we have now it would make it a bit easier, especially the conflict part. We also shaved some 130 lines, which is also good (and I could theoretically shave even more if I weren't to write those extensive javadocs)