Update dependencies & increased switch cooldown
The issue with the Switch enchantment lies in that the event was handled every tick, however sometimes the block was touched with a 1 second delay. Also added another Javadoc.
This only fixes a bug that shouldn't occour usually, but doesn't fix the actual bug (action cooldowns do no longer work)