3.0.0-rc.1 release
I think the plugin is now in a usable state again and I'm kindof in a high pressure
to finish this update. In full it misses one feature I would've liked
to implement, but in retrospec would require an almost full rewrite of the plugin
with a lot of pain

Knownbugs:
  Gust has some issues in combination with sneaking (this is probably intended)

The release condidate stage is meant to patch final issues that arise as well
as making some final adjustments in case I missed a feature that should've been
implemented into it (such as optimisations), so it's not entirely bug free.