Fixed one of the most notorious bugs to exist
I should have implemented this way sooner, but hey, after over 3 months
of asking for this, it has been implemented.

Also reverted a case of API deprecation as it would be just suicidial to
do it now, but it will be done later on when 4.x.x is in development, hopefully.