Implement support for LogBlock
It appears that other logging plugins such as Prism and CoreProtect are
questionable at best. Coreprotect has it's source unavailable
and binding is further impossible as it does not have any maven
artifact that I know of.
Prism in turn does not have the APIs that I need in order for it to
work. (apparently it does not allow me to tell the plugin that something
changed)

Additionally this change only affects a few enchantments:
- Spectral
- Switch
- Terraformer

I also updated the readme file to reflect on the current situation:
- It is unlikely that 4.0.0 will be released anytime soon, let alone
  that it would use paper api natively (I did try this out already,
  it appears that the Paper API has not much to provide outside of
  features that I cannot make use of.)
- Paper is pretty vocal on a hard fork, if that is performed support
  cannot be guaranteed. Even though I use paper myself, I bind with
  spigot. Either way, I wish them good luck for when they actually
  try it.
- I currently have other projects than developing minecraft plugins,
  which I enjoy a bit more. This leads me to neglecting this repo
  (look, it took over a week to ship this feature, and even then it is
  incomplete). As such I cannot really say that the repo is actively
  maintained. That being said I am a bit quicker for bugfixes, provided
  that they are easy to catch.
- I have found out that the GPL needs you to list install instructions,
  so I added them (The GPL is pure nonsense)
- Add a notice over premature optimization. In truth almost all
  optimization were done prematurely here. Except for a few enchantments
  I did not actively benchmark the improvements.

This probably has to be my largest commit message so far, what a waste
of time given that noone reads them (I can't blame anyone).