Crack down on static abuse and fixed enchantment conflicts
Quite a lot of work and in the end I had to also commit a bit of static
abuse, but in the end I removed more than I added.
I also renamed a few variables and methods and fixed enchantment
conflicts after having added the framework for it in 2.1.6 and
realising that there was an issue within 3.0.0 no Idea why that took so
long.

As always, I do not test pretty new branches where there is active
development, so other that this successfully compiles there are not
larger gurantees