Turns out CP actually has a maven release
Still stupid design, why would I be expected to search for "maven" in
their API documentation that is litterally useless for me.
Like in all seriousness, what is "Specify the BlockData of the block you're logging. Can be set to "null"."
Is it the BlockData from before the interaction or the BlockData from
after? Honestly, I don't bother to test